requirements.txt
pyrogram
python-dotenv==0.19.0
requests
pymongo
gpytranslate